Adding decimales to shp and extracting to array

+++ preprocessing/get_temperature.py
@@ -25,7 +25,7 @@ def get_month_band_temperature(layerShp,rlayer):
     if caps & QgsVectorDataProvider.AddAttributes:
         res1=layerShp.dataProvider().addAttributes([QgsField("time_str",QVariant.String)])
         res2=layerShp.dataProvider().addAttributes([QgsField("raster",QVariant.String)])
-        res3=layerShp.dataProvider().addAttributes([QgsField("TEMP",QVariant.Int)])
+        res3=layerShp.dataProvider().addAttributes([QgsField("TEMP",QVariant.Double)])
         res4=layerShp.dataProvider().addAttributes([QgsField("month",QVariant.Int)])
         
     id=0
@@ -64,7 +64,7 @@ def get_month_band_temperature(layerShp,rlayer):
         
         x=long[i]
         y=lat[i]
-        temp = int(get_temperature(rlayer,band_num,x,y,type='geo'))
+        temp = float(get_temperature(rlayer,band_num,x,y,type='geo'))
         temperatures.append(temp)
         rasterband.append(band_num)
         time_encode.append(encode)
